+++ src/solver.jl
@@ -132,7 +132,8 @@ Main solver structure for the Vortex Step Method.See also: [solve](@ref)
 ## Additional settings
 - `type_initial_gamma_distribution`::InitialGammaDistribution = ZEROS: see: [InitialGammaDistribution](@ref)
 - `use_gamma_prev`::Bool = true: reuse provided previous gamma as initial guess when available
-- `core_radius_fraction`::Float64 = 0.05:
+- `core_radius_fraction`::Float64 = 0.05: Bound vortex core cut-off, as a fraction of the
+    filament length, following Damiani et al. (2019)
 - mu::Float64 = 1.81e-5: Dynamic viscosity [N·s/m²]
 - `is_only_f_and_gamma_output`::Bool = false: Whether to only output f and gamma
 - `reference_point`::MVec3 = [0.0, 0.0, 0.0]: Moment reference point in body frame